Fix fragile test in ParameterizedTestIntegrationTests
Prior to this commit, the executesLifecycleMethods() test in
ParameterizedTestIntegrationTests would periodically fail due to the
fact that Java does not guarantee the order in which methods are
returned when looked up via reflection.

This commit fixes this fragile test by tracking the order in which the
individual test methods were executed and building the resulting
expectations dynamically.