Revise Assertions.assertAll() to track all exceptions
Prior to this commit, Assertions.assertAll() only tracked exceptions of
type AssertionError; however, this behavior was unintuitive.

This commit therefore revises Assertions.assertAll() so that it tracks
exceptions of any type, unless the exception is a blacklisted exception
in which case it will be immediately rethrown.

As an added bonus, the performance of Assertions.assertAll() has been
slightly improved since a MultipleFailuresError is now only
instantiated when necessary.

Issue: #808