Add DynamicContainer support
Prior to this commit only a flat collection of dynamically generated
tests was allowed as the return type of a @TestFactory annotated
method. The collection entry type was set to `DynamicTest`.

This commit introduces the abstract base class `DynamicNode` and a
`DynamicContainer` class collecting node instance. @TestFactory
annotated methods are now allowed to return a collection of nodes --
which can be a test or a named container of tests.
Creating a dynamic tree of tests like that gives users the ability to
better structure the generated tests.

The old behaviour, returning only instances of `DynamicTest`, is preserved.