Improve display name for array args in parameterized tests
Prior to this commit, a display name for a @ParameterizedTest that
accepted an array as an argument would result in the array being
displayed as `[Ljava.lang.String` which is not very user friendly.

This commit addresses this issue by converting array arguments for such
tests to a human readable format — for example: `[foo, bar]`.

In addition, this commit moves the existing array formatting code from
ToStringBuilder to a new StringUtils.nullSafeToString(Object) method.
Consequently, the existing nullSafeToString(...) methods in StringUtils
that deal with classes have been moved to a new dedicated ClassUtils
class in order to avoid ambiguity issues due to overloaded methods and
varargs.

Issue: #814