Don't enforce syntax highlighting for markdown.
The document is not rendered completely in my IDE when the part for
the syntax highlightning is present. With a simple code markup it is
rendered completely. GitHub renders itself already renders the snippet
without syntax highlightning. This means this commit does not change
the rendering of the page on github.com.