Rename method to getFilterByType.
Provide a consistent API. The method for retrieving selectors already
has the name `getSelectorsByType` instead of
`getDiscoverySelectorsByType`. I think it is helpful to omit the word
discovery in the method name because it reduces its length but the
reader still can deduce from the context (it's an
`EngineDiscoveryRequest`) that the methods return `DiscoveryFilter`s and
`DiscoverySelector`s.