Find annotations on interfaces
Prior to this commit, the findAnnotation() methods in AnnotationUtils
did not search on interfaces; however, the findRepeatableAnnotations()
methods did search on interfaces.

This commit addresses this discrepancy by searching for annotations on
interfaces within findAnnotation() before searching for indirectly
present annotations (if the supplied element is a class).

Consequently, annotations such as @TestInstance can now be declared
on test interfaces (a.k.a., testing traits).

Issue: #907