Introduce and enforce syntax for tags
Prior to this commit, tags were trimmed (#942) but were otherwise
allowed to contain any type of character, including whitespace or ISO
control characters. However, this degree of flexibility can lead to
subtle configuration errors and also makes it challenging to reason
about a tag expression language as proposed in #927.

This commit addresses these issues by introducing and enforcing a
syntax for tags. Specifically, all tags must now conform to the
following syntax rules.

* A tag must not be null or blank.
* A trimmed tag must not contain whitespace.
* A trimmed tag must not contain ISO control characters.

Consequently, the following changes make it impossible to create a
TestTag or TagFilter based on a tag that is not syntactically valid.

* The TagFilter.includeTags(), TagFilter.excludeTags(), and
  TestTag.create() factory methods now throw a
  PreconditionViolationException if a supplied tag is not syntactically
  valid.
* @Tag declarations containing invalid tag syntax will now be logged
  as a warning but effectively ignored.

Furthermore, a new TestTag.isValid(String) method has been introduced
for checking if a tag is syntactically valid. This allows test engines
to preemptively validate the syntax of a tag provided by a user before
attempting to register such an invalid tag.

Issue: #956