Improve performance of test discovery in Jupiter & Vintage
Prior to this commit, Jupiter and Vintage both executed exhaustive
searches across the entire class hierarchy in search of all methods
annotated with @Test, just to determine if there was a single such
method.

This commit addresses this issue by introducing a new isMethodPresent()
method in ReflectionUtils that searches for a method based on a
supplied predicate and returns true as soon as such a method is found.

Issue: #983