Catch InternalError only for specific code
The InternalError is normally handled like any other Throwable except
when its cause is a malformed class name. The only statement that may
throw such an error is the statement that loads the class. Therefore
the exceptions/errors thrown by other statements can be handled by
generic catch block for Throwables without changing the behaviour of
the method.