Log additional information about partially resolved Unique IDs
Prior to this commit, if a selected Unique ID could not be fully
resolved in JUnit Jupiter, the JavaElementsResolver logged a warning
stating that the Unique ID could not be resolved at all. This was,
however, extremely misleading if the Unique ID was in fact partially
resolved.

This commit addresses this issue by providing additional information
in the log message if a Unique ID can only be partially resolved.

Issue: #1031