Fix bug in FilePosition.equals(Object)
Prior to this commit, the equals() implementation in FilePosition
compared the column numbers using '=='. This worked fine when the
field was an 'int', but with the recent change to 'Integer' for the
field type, this comparison no longer worked.

This commit fixes this issue by using Objects.equals() to compare the
column numbers.

Issue: #939