Remove compile dependency on apiguardian-api in Maven POMs
Prior to this commit, published artifacts contained a `compile` scope
dependency on `apiguardian-api` even though no runtime dependency
actually exists, and this leads to problems in certain build scenarios.

This commit addresses this issue by switching from Gradle's `api`
configuration to its `compileOnly` configuration. Note, however, that
this completely removes the dependency on `apiguardian-api` from the
generated Maven POMs.

Ideally the dependency on `apiguardian-api` should be published as an
`optional` dependency in the generated Maven POMs, and we will look
into a solution to achieve that in a later commit.

Issue: #1065