Upgrade to JRuby 9.1.13.0 and build script cleanup
Prior to this commit the JRuby version bundle in the AsciiDoctorJ Gradle
plugin was used: 9.1.12.0 -- which contains a bug that prevent the PDF
backend to work properly on Windows 10.
The manual upgrade can be removed when AsciiDoctorJ defaults to this or
a newer JRuby version.

Removed "standalone" configuration, the `testRuntimeClasspath` was
already configured to run the `ConsoleLauncher --help` successfully.
Using java.nio.file.Path(s) where possible and store simple String in
the ext declaration.