Do not silently swallow exceptions in ExpectedExceptionSupport
WARNING: this commit fixes the original bug but introduces a new one!

The cause of the original bug is that
`AbstractTestRuleSupport.handleTestExecutionException()` effectively
swallows any exception if there are no registered instances of the
`ExpectedException` rule.

This commit fixes the original bug by counting the number of _members_
that are discovered by having
`invokeAppropriateMethodOnRuleAnnotatedMembers()` return the number of
members and rethrowing the exception if the number of members is zero.
However, this change introduces a new bug since
`ExpectedExceptionSupport.handleTestExecutionException()` invokes
`AbstractTestRuleSupport.handleTestExecutionException()` twice: once
for fields and once for methods. Since the "member count" does not
currently aggregate across both fields and methods, the counting logic
is fundamentally broken.

To fix the new bug introduced in this commit, we will therefore need to
count the number of all members registered; otherwise, we don't know if
the adapter is supposed to rethrow the exception.

Issue: #1069