Use module descriptor for artifact and version information retrieval
Prior to this commit the default implementation for extracting the
artifact ID and version of a test engine instance used only the package-
related attributes in the manifest of the JAR file.

Now, the default implementation checks the Java Platform Module System
for the those information -- when running on Java 9 or higher.

Closes #600