Make custom logging façade testable
This commit introduces a LogRecordListener that can be registered with
JUnit's logging façade for JUL via the LoggerFactory. Such a listener
can be queried for all log records that were submitted since the
listener was registered. This makes it possible for tests to register
a listener for a given scenario that is supposed to log something and
then verify that the expected message got logged at the expected level.

After such a test scenario has completed, the test code should remove
the listener from the LoggerFactory.

To simplify matters, the creation of a listener, registration of a
listener, and removal of a listener are all encapsulated in a JUnit
Jupiter extension that is registered via a new @TrackLogRecords in the
junit-jupiter-engine test source tree.

See MultipleTestableAnnotationsTests for an example of this
functionality in action.

Issue: #1047