Generate default failure messages for assertTrue() & assertFalse()
Prior to this commit, assertTrue() and assertFalse() in the JUnit
Jupiter Assertions class did not generate default failure messages.

This commit improves the usability of such failure scenarios by
generating meaningful default failure messages for assertTrue()
and assertFalse().

Failure messages for assertTrue() and assertFalse() are now aligned
with failure messages for assertEquals.

For example, the generated failure message for a call to
`assertTrue(false)` is now "expected: <true> but was: <false>" instead
of an empty string.

Similarly, a call to `assertTrue(false, "custom")` now results
in a failure message of "custom ==> expected: <true> but was: <false>".

The failure message generation for assertFalse() is analogous to the
new behavior for assertTrue().

Issue: #1095