Align Vintage and Jupiter version numbers
Using a version for Vintage that aligns with JUnit 4 seemed a good idea at first. However, it has proven to be impractical to keep up for two reasons. A minor release of Vintage would have to be 4.13, even when there's no JUnit 4.13, yet. 4.13.1.0 does not work because Maven sorts version numbers with four parts alphabetically. Thus, we have decided to align the version number of Vintage with Jupiter for now.