Use current module layer when looking for test classes
Prior to this commit, only the boot module layer was used when scanning
resolved modules for test classes. Now, try current module layer before
resorting to the boot layer as a fall-back. The current module layer is
the one that the class loader of class `ModuleUtils` returns:

  var layer = ModuleUtils.class.getModule().getLayer()

This commit also adds more debug logging statements.

Addresses #1175