Introduce Details.SUMMARY and mute output in mode NONE
Prior to this commit a summary table of counts was printed even when the
details mode NONE was passed to the console launcher. Now, when mode
NONE is selected, nothing is printed. Unless a failure was recorded.

The new SUMMARY mode prints the same output that was printed before when
NONE was selected.

The default details mode value of the Gradle Plugin remains NONE. If you
want to see the summary table emitted you now need to configure it like:

junitPlatform {
	details 'summary'
	filters {
		engines {
			include 'junit-jupiter'
		}
[...]
}

Closes #1179