Stop closing EngineExecutionContext in Node.cleanUp() by default
Prior to this commit, the execution context was closed by each Node
in cleanUp() regardless whether it actually created a fresh instance in
prepare(). Since the default implementation of prepare() returns the
supplied execution context unmodified, the default implementation of
cleanUp() must not close it.

Thus, this commit removes close() method in EngineExecutionContext.
In addition, a bug in the Jupiter engine is fixed, where dynamic nodes
closed their parent execution context.

Issue: #742