Refactor module-path support in Gradle Plugin
Prior to this commit there was simple boolean flag that triggered the
use of the JPMS module-path. The flag is now replaced by an empty file
collection that can filled with module-path entries at configuration
time.

If a non-empty modulepath collection is present the following JVM
arguments are added to the JavaExec task, unless already specified:

java
  --module-path <modulepath>
  --add-modules ALL-MODULE-PATH
--module
  org.junit.platform.console
  --scan-modules

Examples usage from a 'build.gradle':

junitPlatform {
	modulepath = files(configurations.testRuntimeClasspath, compileTestJava.destinationDir)
	details 'tree'
}