Introduce ScriptEvaluationException
Prior to this commit a generic JUnitException was thrown when an error
occurred while evaluating condition scripts. Now an instance of the
specialized ScriptEvaluationException is thrown signalling the context
of the error.

This commit also treats `null` results returned from scripts as an
error: `null` can't be converted to an unambiguous instance of
ConditionEvaluationResult.