Export `findAllClassesInModule` in `ReflectionSupport`
Prior to this commit, 3rd-party engine implementors had to use internal
classes to achieve the same behaviour as the JUnit Platform when
scanning for classes in modules. Now, the findAllClassesInModule method
is officially supported.

This commit also changes direct usages of `findAllClassesInModule` from
`ModuleUtils` to the new method declared in `ReflectionUtils` within
the codebase, because `ModuleUtils` is just a multi-version overlay of
reflection-based utilities.

Closes #1300