Introduce @Nested test with annotated constructor parameter
This commit introduces a @Nested test to verify support for constructor
injection via a ParameterResolver that looks up an annotation on the
Parameter supplied via the ParameterContext.

The test is @DisabledOnJre(JAVA_8) due to a bug in javac that was first
fixed in JDK 9. Note, however, that the test executes and passes on both
JDK 9 and JDK 10.

Issue: #1345