Introduce annotation lookup methods in ParameterContext
Due to a bug in javac on JDK versions prior to JDK 9, looking up
annotations directly on a java.lang.reflect.Parameter will fail for
inner class constructors.

Specifically, the parameter annotations array in the compiled byte code
for the user's test class excludes an entry for the implicit enclosing
instance parameter for an inner class constructor.

This commit introduces a workaround for this off-by-one error by
helping extension authors to access annotations on the preceding
Parameter object (i.e., index - 1).

This workaround is made available to extension authors via the
following new convenience methods in the ParameterContext API.

- boolean isAnnotated(Class<? extends Annotation> annotationType);
- Optional<A> findAnnotation(Class<A> annotationType);
- List<A> findRepeatableAnnotations(Class<A> annotationType);

Issue: #1345