Support "Test*" prefix for test class names by default
Prior to this commit, the JUnit Platform only supported "*Test" and
"*Tests" suffixes for class names in the default include pattern.

This commit modifies the regular expression used in the default pattern
to support a "Test*" prefix as well.

This pattern is used by the ConsoleLauncher, the JUnit Platform Gradle
Plugin, and the JUnitPlatform runner.

Issue: #1354