Ensure Ant tests pass when ANT_HOME is set locally
Prior to this commit, if the ANT_HOME environment variable was already
set on the local machine, the Ant tooling support integration tests
would use the local Ant installation instead of the one downloaded by
the tooling support test infrastructure.

This commit addresses this issue by explicitly setting the ANT_HOME
environment variable to the location of the downloaded Ant binaries.

Issue: #1429