Require JDK 10 for Eclipse usage
Prior to this commit, all JUnit 5 projects could be imported into
Eclipse using JDK 8 (except the junit-platform-commons-java-9 project
which naturally required JDK 9). The recent introduction of the
platform-tooling-support-tests project requires that the version of
Eclipse in use support JDK 10.

This commit therefore removes the previous custom Gradle configuration
for supporting a mixture of JDK 8 and 9, thereby requiring that
developers use a version of Eclipse IDE that supports JDK 10 (e.g.,
Eclipse 4.7.3a or higher).

Consequently, developers must be careful not to introduce dependencies
on JDK 9 or 10 libraries or language features while developing within
Eclipse. Note, however, that such usage should cause the Gradle build
to fail in any case.