Add --fail-if-no-tests command line option to ConsoleLauncher (#1399)
Prior to this commit, if the ConsoleLauncher found no tests it would
exit with a success status. This could happen by user mistake -- for
example, due to improper use of a selector. In such situations, a user
would not be informed of the misconfiguration.

This commit introduces a new `--fail-if-not-tests` command-line option
for the ConsoleLauncher. This flag can be set to `true` to make the
launcher fail when no tests are found, helping users catch such
misconfigurations.

Closes: #1298