Improve error messages for ParameterResolutionExceptions
Prior to this commit, error messages for ParameterResolutionExceptions
generically referred to a constructor or a method as an "executable".

Although that is technically correct, many developers are not familiar
with the fact that Constructor and Method share Executable as a common
superclass. Consequently, the term "executable" is somewhat cryptic for
most developers.

This commit addresses this issue by explicitly stating that the
Executable is either a "constructor" or a "method" when generating
error messages for ParameterResolutionExceptions.

Issue: #1448