Use same JDK as current tests do
Prior to this commit the environment variable JAVA_HOME was passed
untouched to the external tool execution environment. As that
variable might point a different JDK/JRE installation than the
current one executing the tests it could lead to unexpected test
results.

Now, the path for JAVA_HOME is extracted from current process by
default. If an tests needs a dedicated JDK/JRE installation, it
is free to pass it via its environment map.