Introduce support for parallel test execution
This commit adds opt-in support for parallel test execution and
capturing output to `System.out` and `System.err`. Both features are
disabled by default but can be enabled and configured using
configuration parameters.

The implementation is based on the Fork/Join Framework and designed to
be reusable by other test engines that extend HierarchicalTestEngine.

The Jupiter API provides annotations to declare which shared resources a
test needs to access and in which way. Moreover, the execution mode of a
test can be influenced.

In addition, a number of TestExecutionListeners have been made
thread-safe.

The documentation subproject is now configured to execute tests in
parallel. All other subprojects will have to wait as Gradle currently
blows up when used with parallel test execution.

Resolves #60. Closes #1461.

Co-authored-by: Leonard Brünings <lord_damokles@gmx.net>
Co-authored-by: Christian Stein <sormuras@gmail.com>