Resolve TestInstanceFactory at class level
Prior to this commit, TestInstanceFactory extensions were resolved
lazily on-demand. This led to the undesired side effect that
configuration errors were reported for each test method when executing
with per-method lifecycle semantics. In addition, this also meant that
an attempt was made to invoke all test methods even though the test
class could not be instantiated.

This commit addresses these issues by resolving the TestInstanceFactory
during the "prepare" phase for each ClassTestDescriptor.

This commit also introduces the following methods in ExtensionRegistry
in order to make the above change possible.

 - getParent()
 - getLocalExtensions(Class)

Issue: #672