Allow Assertions and Assumptions classes to be extended
Prior to this commit, it was not possible to extend the Assertions and
Assumptions classes since they both were final and had private
constructors.

This commit removes this restriction in order to support unusual
circumstances in which it may be "forbidden" to use static imports.

Note, however, that a note has been added to the class-level JavaDoc for
these classes in order to voice the opinion of the JUnit Team that it is
a best practice to use static imports for methods defined in the
Assertions and Assumptions classes.

Resolves: #1484