Safely parse FilePosition from query string
Prior to this commit, FilePosition.fromQuery() would fail (and return
an empty Optional) if any query parameter had a non-integer value.

This commit fixes this issue by only attempting to parse a query
parameter's value as an integer if the parameter's name is "line" or
"column".

In addition, the parsing algorithm now stops as soon as it has
encountered entries for the line and column.

Fixes: #1489