Avoid redundant type argument warning on JDK 9+
This commit introduces a concrete implementation of the internal
ResultAwareThrowingSupplierAdapter API in order to avoid a warning
that results from the use of an anonymous inner class with "redundant
type information" that is flagged on JDK 9 and above. Since the code
base is still released using JDK 8 as the target environment, we cannot
simply use the diamond notation for the anonymous inner class
declaration.