Add openjdk11 axis to the build matrix
With `openjdk-ea` now pointing to http://jdk.java.net/12 we need
to explicitly include `openjdk-11` into the Travis CI build matrix.

The fix for the NPE in the TLS 1.3 implementation is available
with build 22 on jdk.java.net/11 -- thus the build should work now
on JDK 11. Leaving the axis in the `allow_failures` group until
JDK 11 is promoted to GA.

Issue: #1487