Handle exceptions in StringUtils.nullSafeToString()
Prior to this commit, if the toString() implementation of an object
supplied to nullSafeToString() threw an exception, that exception would
simply be allowed to propagate. Consequently, unexpected failures could
occur simply because the framework was attempting to safely generate a
String representation of a given object.

This commit improves the "safety" of nullSafeToString() by catching any
exception encountered while attempting to safely generate a String
representation. If an exception is caught in the process,
nullSafeToString() will now generate a default String representation
based on the object's fully qualified class name and system hash code,
separated by an "@" symbol.

This change affects the following use cases.

- The generation of detailed failure messages for failed assertions no
  longer fails if the toString() implementation of an object supplied to
  the assertion throws an exception.
- Display name generation for a @ParameterizedTest no longer fails if
  the toString() implementation of an argument for the parameterized
  test throws an exception.
- The use of ToStringBuilder for building strings is now safer.

Issue: #1492