Use correct class for method-based test descriptors
Prior to this commit, `MethodBasedTestDescriptor` and
`VintageTestDescriptor` did not preserve the original test class of the
discovered method and instead obtained it from the
`java.lang.reflect.Method`. This was problematic for inherited test
methods, since they appeared to come from the declaring class instead of
the current test class. In addition, the Maven Surefire provider failed
to match such methods when executing tests matching a mask (e.g., via
`-Dtest=*MaskTest`).

This commit fixes this problem by modifying `MethodBasedTestDescriptor`
and `VintageTestDescriptor` to ensure that the underlying
`MethodSource` preserves the correct, current test class.

Fixes: #1406