Fix ClassSource implementation
Prior to this commit, ClassSource did not enforce the precondition for
a class name. In addition, equals() and hashCode() were incorrectly
based on the optional class instead of the required class name.

This commit fixes both of these sets of bugs and introduces proper unit
tests for all preconditions and features.

Fixes: #1506