Increase parallelism of dynamic test execution
This commit increases the amount of allowed queued work for the worker
thread to the desired level of parallelism. This optimistically assumes
that the already queued tasks can be stolen by other workers and the new
task requires about the same execution time as the already queued tasks.
If the other workers are busy, the parallelism is already at its desired
level. If all already queued tasks can be stolen by otherwise idle
workers and the new task takes significantly longer, parallelism will
drop. However, that only happens if the enclosing test task is the only
one remaining which should rarely be the case.

Issue: #1491