Implement equals() & hashCode() in DiscoverySelector impls
Due to an oversight of the team, implementations of DiscoverySelector
did not have custom implementations of equals() and hashCode() prior to
this commit. This is in contrast to implementations of TestSource,
TestDescriptor, etc.

This commit addresses this issue by implementing and testing equals()
and hashCode() for all concrete DiscoverySelector implementations.

Fixes: #1510