Add TestReporter::publishMessage
So far, the TestReporter only offered the publication of key/value
pairs (with `publishEntry`), which is unsuitable for simple status
messages as it is not clear what the key would be. This commit adds
a method `reportMessage(String)` that requires no key.

Internally `publishMessage` calls `publishEntry` and uses the key
"message", but that is regarded as an implementation detail.

Since `TestReporter` is essentially just a specialized facade for
the `ExtensionContext`, a corresponding  method
`publishReportMessage` was added there as well.