Ensure custom ClassLoader for assertThrows() executes on JDK 9+
Prior to this commit, the EnigmaClassLoader used in
AssertThrowsAssertionsTests only ran on JDK 8. Since the build runs on
JDK 10, however, that means the test for an exception thrown by a
custom ClassLoader has not been executed in the build in quite some
time.

This commit addresses this issue by modifying the EnigmaClassLoader so
that it no longer assumes the system ClassLoader is a URLClassLoader
(which is no longer the case on JDK 9 and above).