Fixing concurrent access issue in TreePrintingListener
The addNode operation in the TreePrintingListener was non-atomic and
subject, therefore, to a race condition. This change makes the methods
that access the map synchronized to remove that race condition. Since
the access of the map is locked on the "this" object by the synchronized
keyword, the use of the ConcurrentHashMap is no longer needed and is
therefore an extra set of locking that is unnecessary.

Issue: #1524