Rename publishMessage to publishEntry in TestReporter API
This commit renames publishMessage() to publishEntry() in the
TestReporter API in order to align with the naming convention and
semantics of the existing publishEntry() variants.

This commit also renames publishReportMessage() to publishReportEntry()
in the ExtensionContext API for analogous reasons.

Furthermore, the default key has been renamed from "message" to "value"
in both cases.

Issue: #1515