Making locking a private implementation detail of TreePrintingListener
The fact that the map needs to lock is really a private implementation detail
that external classes don't need to know about and, if someone were to lock on
the listener externally for some reason, it could lead to a deadlock. Making the
lock be on an explicit internal object rather than the "this" object.

Issue: #1524