Document how to register static fields in Kotlin
JUnit 5 extensions can be registered programmatically.  When one wants to
register extensions that support lifecycle methods, they should be registered
as static fields.

In the Kotlin programming language this is normally achieved using the
`@JvmStatic` annotation.  However that will not work because a private field
will be generated, and private fields are filtered out by JUnit 5.  The correct
approach is to use `@JvmField` instead.

This commit documents the correct approach and provides a Kotlin example
analogous to the existing Java one.

Reference: #1500