Document registering extensions via static fields in Kotlin
JUnit Jupiter extensions can be registered programmatically using
`@RegisterExtension`. When one wants to register extensions that
implement class-level extension APIs, they have to be registered as
static fields.

In the Kotlin programming language this is normally achieved using the
`@JvmStatic` annotation.  However, that will not work because a private
field will be generated, and private fields are filtered out by JUnit
Jupiter. The correct approach is to use `@JvmField` instead.

This commit documents the correct approach and provides a Kotlin example
analogous to the existing Java one.

Resolves #1500.