Fix TestInstanceFactory support for @Nested test classes
Prior to this commit, a TestInstanceFactory registered on an enclosing
class was not inherited by the enclosed @Nested test class. In
addition, if the same type of TestInstanceFactory was registered on
both a @Nested class and one of its enclosing classes, the factory was
completely ignored.

This commit fixes both of these issues.

Fixes: #1567