Further reduce memory footprint of parallel execution
Prior to this commit, the complete `TestDescriptor` tree of a
`TestEngine` that extends `HierarchicalTestEngine` (such as the Jupiter
engine) was wrapped in `NodeTestTask` instances prior to its execution.
This was done in order to store the `ResourceLock` and the forced
`ExecutionMode` as determined by the `NodeTestTaskWalker` class (in
order to ensure deadlock-free execution).

Now, the (renamed) `NodeTreeWalker` returns a `NodeExecutionAdvisor`
which stores this information and provides it to `NodeTestTask`
instances. Thus, `NodeTestTask` instances are created on the fly instead
of up front.

In addition, the new `NodeTestTaskContext` class provides access to
`EngineExecutionListener`, `HierarchicalTestExecutorService`,
`ThrowableCollector.Factory`, and `NodeExecutionAdvisor` which are
identical for all `NodeTestTasks`. Previously, each `NodeTestTask`
had separate references to the first three of these.

Resolves #1578.